Fixes and small additions on map v2 (#156)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

- Removed invisible tables in the first private room at the inn's baths.
Added normal fancy tables instead
- Town house across the merchant's shop now has a normal working door,
keys now fit
- Added a table and stools to the mansion's basement for guards to keep
watch on noble prisoners and the vault. Provided them with playing cards
to kill time
- King now has access to his mansion's telescope, go find it. Also, one
of the rooms in his quarters was repurposed to a personal reading room
- Added a bit of furniture to bog gates, training dummies included.
- Added more torches to the farm, per request of farm mains
- Added a door to the second floor in the farmstead, uses farm key to
unlock.

## Why It's Good For The Game

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

Improving map with quality of life changes. Fixing **very** old mapping
mistakes with the invisible tables and not fitting keys to buildings
they spawn at.