Fixes fireball hitbox (#185)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Fixes https://github.com/Rotwood-Vale/Ratwood-Keep/issues/123 ?

There is no targetting issues with fireball
Fireball has no homing or entity seeking
The fireball used to have a 3x3 hitbox, and this PR makes it 1x1

Also makes other "area" projectiles default to 1x1 range, whichever that
could be, cause I dont think the intention was to ever increase the
hitbox

Spent goddamned 2 hours debugging and trying to see what's wrong with
targetting or homing, but it was neither of that

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->