Heartfelt & Witch Hunter subclasses: allow tolerated+ races (#181)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Changes Lord of Heartfelt, Hand of Heartfelt, and Witch Hunter to no
longer be humen exclusive. Tested in-game before and after the change.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
If King and other nobles can be whatever race aside from horc, kobold or
goblin, why can't the Heartfelters? Same goes for the Inquisitor and the
Witch Hunter, where both are practically identical.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

---------

Co-authored-by: zoktiik <zoktiik@users.noreply.github.com>