"allowed_patrons" code now tries to avoid switching faiths first (#190)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
The logic used is a bit different than the title implies, but the effect
is the same.
Jobs and subclasses that switch your patron when you choose them will
now attempt to avoid patrons that are from a different faith, instead
only picking a patron randomly from within the same faith if possible,
then falling back to random selection.

Old God ("One-Psydonite") characters becoming an unsupported (sub)class,
that _does_ support both Divine and Inhumen patrons, will randomly pick
between them as normal; there is no special logic or cases here.

With this PR, an Astratite character will only ever become a Pestrite
barber-surgeon or a Ravoxian barbarian, never a Graggite (if possible).
This does not prevent him from becoming a Psydonite witch hunter or a
Zizoist necromancer.

In other words, this PR really only means that you'll be seeing a lot
more of this:


![image](https://github.com/Rotwood-Vale/Ratwood-Keep/assets/50035052/7fb7b190-db40-4c85-95ef-344635dc5596)

![image](https://github.com/Rotwood-Vale/Ratwood-Keep/assets/50035052/1f52fedf-f7b7-4d8a-9aff-c8aa3fb79d66)

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Having to roll the dice on whether my character is switched to a
different faith is bad.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

Co-authored-by: zoktiik <zoktiik@users.noreply.github.com>