Maybe fixes zombie decap bug (#162)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Currently there is a bug where a decapitated zombie will ghost the
player whose body the zombie had occupied, even if the player respawned
and is playing in a new body

Maybe this fixes it. I'm unsure
We can't remove the mind because minds have skills information


<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->