Adds Swept Harpy Wings. (#212)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Title
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
They look good +++ current harpy wings kinda suck.

Also looks good for Eastern Dragons (the entire reason I even decided to
port them).

This will likely become coal because they hide the sides of your
character a bit too well but so do Drakian wings.


![image](https://github.com/Rotwood-Vale/Ratwood-Keep/assets/65188584/60fcaed4-b7ae-47e2-94b1-b3ab4814fba4)

![image](https://github.com/Rotwood-Vale/Ratwood-Keep/assets/65188584/dc7ae059-f8ae-420b-afa2-b0a5c0cfea82)

![image](https://github.com/Rotwood-Vale/Ratwood-Keep/assets/65188584/58a8cd5f-8026-48a7-94ec-53bc4b5f4dad)


Tentatively, credits to SPLURT (they didn't sprite this but I can't find
a source for the sprite and they had it) for the icons.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

---------

Co-authored-by: Azarak <azarak10@gmail.com>