Boost Maniac's Medicine skill to 3 for surgery if needed (#203)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
**EDIT:** Azarak kindly provided code to only boost the Maniac's
medicine to 3 only if needed, and without adding it on top of the
previous skill. Thoroughly tested the code, ready for merge.

Old description:
The grug solution of resolving issue #202. This will make sure the
Maniac always has Medicine rank of 3 or higher by tacking it on top of
his existing skill, which acts as a somewhat insane buff.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
The Maniac is able to accomplish his goals with the tools he's supposed
to be able to use.
I don't have to write a proc that actually touches the XP system.
As to whether it's balanced, uh, I dunno. YMMV. Works on my machine. Up
to you bud.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

---------

Co-authored-by: zoktiik <zoktiik@users.noreply.github.com>
Co-authored-by: Azarak <azarak10@gmail.com>