disables pq gain for burial because morticians keep stealing perfectly good bodies to bury for pq, doubles reward for coining (#232)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

morticians keep doing this shit to the point that i have seen morticians
drag bodies that are IN THE PROCESS of being revived to the burial site
for their pq. in fact i have seen them do it to LIVING PEOPLE for pq, so
fuck you, no more getting pq for round removing people. you still get pq
for giving a coin - double now in fact - because that actually HELPS the
victim after death: burial does NOTHING

---------

Co-authored-by: oilysnake <63020759+oilysnake@users.noreply.github.com>