Rebels victory round end changes (#215)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Rebels now need to sit on the throne for 5 minutes (30% faster for every
other nearby rebel) to be able to declare victory.

Rebel victory now will will either:
1. End the round in 15 minutes if the round passed 2:30hr mark
2. End the round at the 2:30hr mark if the round did not pass it.

PR is actually not tested cause testing gamemode stuff is hell lmao

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->