Hear() now gets the original message & King commands fixes (#239)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Hear() now receives the original_message that does not get treated by
any slurring or accents (despite already getting "raw_message", which
ends up to not really being as raw as you might think)

Also fixes one case of instant rebel victory due to copypasta code that
existed there before

The Throne commands now use the original message to recognize commands,
and to get names of outlaws.
The contents of laws, decrees and announcements will still be treated
and accented

![dreamseeker_UBo6uYVhYX](https://github.com/Rotwood-Vale/Ratwood-Keep/assets/6381979/5c5d4f80-c607-4afc-a3b7-fbaa49276589)



<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->