Immediate say & me. Big me has seperate keybind (#234)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
The me verb (keybind M) now is back to small input text, with enter
sending the message.
There's a new verb - me(big) (keybind ,) which is a `me` with the big
message input

The say, me and big me verbs now have immediate input and will not wait
for server to prompt the client the input box
Means no more lag before it appears after pressing T or M

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->