Customization updates (#244)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Removes duplicate bunny ears
Adds new perky bunny ears (credits to Skyrat)
Makes it so the Tail and Tail Feature offsets are based on underwear
offsets, and not back offsets, resulting in 1 pixel higher tails for
most females
Updates underwear offsets to be proper for dwarves, kobolds and
verminvolk
Allows Halfkin socks, tips markings customizations
Allows Halfkin much more tails and ears customizations
Makes it so Halfkin use another body sprite that is awfully similar to
the human sprite but not quite the same (because markings were modeled
after the tiefling body, and it turns out it slightly is different from
the human one) - you can barely tell the difference tbh.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->