Add red wizard robe to crafting + default robe list, give wizards shirts under their robes (#211)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

Orange wizard robes are kinda fugly and don't work with the new red
hats, so I'm adding red as an option/default option.

Also, wizards are topless under those robes and it's kinda weird -
especially if playing a female character because the big booba robe
sprite makes your fucking cleavage hang out (VERY noticable when colors
are different) and it's kinda gross. Gave them shirts under it so it
looks a little better, it covers it up a bit!

![thewizard](https://github.com/Rotwood-Vale/Ratwood-Keep/assets/11761309/1d21722d-1eab-4973-99d2-b58c0f12b2fe)

---------

Co-authored-by: oilysnake <63020759+oilysnake@users.noreply.github.com>