Flat chest doesn't get the extra overlay (#254)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Females which lack boobs, or have flat boobs will not get the extra
overlay meant to represent that bump in clothings, such as shirts, robes
or armors

Mind you, not every clothing was made with the overlay seperation in
mind, so it won't always work as intended

Fixes https://github.com/Rotwood-Vale/Ratwood-Keep/issues/219 - However
due to the difference in sprites of male/female bodytypes it does not
change the clothing variation to male gender, as the issue suggests, but
just doesn't apply that extra overlay.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->