Reverts fishing minigame (#289)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Reverts https://github.com/Rotwood-Vale/Ratwood-Keep/pull/155
It's really menial and annoying.
There is a good bunch of changes bundles up with it but most of them are
really unnessecary
The fix for icon state of a bait bundle is left
And fish rod having the cast line as first intent is left too

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->