[RESPRITE] Reskins "Love Potion" (#271)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## Replaces "Love Potion" sprite

All I did was take the bottles we currenty use, changed it pink and
submitted it into the cooking.dmi

The "Love Potion" is just an item and not a reagent. Perhaps in future
it will be turned into a reagent so wizard may practice alchemy and
spice up the life of the court.


Original Sprite

![image](https://github.com/Rotwood-Vale/Ratwood-Keep/assets/53943894/70c74f76-51cc-47ca-80f9-29fe9afd113e)


New Sprite

![image](https://github.com/Rotwood-Vale/Ratwood-Keep/assets/53943894/a9894648-ce9b-41ab-b890-f9a17f24b504)