Bandit intro popup and combat track change (#277)
## About The Pull Request

Bandit introduction text is now a popup.
Combat track was reverted back to the Evanora Unlimited - Mithlond
(Instrumental), compressed, and also lowered in base decibel so it's not
the earrape it originally was.

## Why It's Good For The Game

Bandits almost always miss the introduction text explaining things,
including their goal, and the idol. This makes sure you can't miss it.
The combat track I never liked, but not many complain, so I guess it's a
matter of taste. Toss alternatives in here if you have any.