fermented liquids are no longer automatically watered down  (#297)
for some reason fermented liquids were being automatically watered down
for... some reason? 3u of water with each crop added

that made it so that all the ale looked like piss and all the wine was
this weird pink color

yeah that's gone now. no reason for it to exist. fermented liquids
provide hydration on their own so