Grenzelhoft's new look and changes (#316)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Bring grenzelhoft drips from the codebase Openkeep to make them more
unique
Along with that, replaced lit torches with torches on their belt due to
runtime and gave grenzel merc key because it is dumb only desert riders
get them


Credit to the one that sprited it because it look actually amazing, I
wasn't able to find them.
https://github.com/Darkrp-community/OpenKeep

also i wasn't able to fix that nanotrasen and syndicate svg issue, it
keep appearing frequently?
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

![image](https://github.com/Rotwood-Vale/Ratwood-Keep/assets/61919894/d0c62f31-264c-44bb-a6f5-a38aefdae0c6)

Drip is always appreciated.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->