House staff reorganization (#187)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

Reduce the butler/maid job slots to 1.
Increase the servant job slots to 6.
Rewrite both job descriptions to clarify the hierarchy between the two.

## Why It's Good For The Game

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

Maids (3 slots) and Servants (4 slots) constituted a redundancy, with
the only difference between them being the youngfolk age lock, which is
meaningless. I've made it so the Maid is considered the head of the
house staff and essentially a majordomo who commands the servants.