Lowers small log stockpile price (#396)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Lowers it from 3 to 2, adjusting other prices accordingly

Problem not exactly solved yet, as what is really the issue is how easy
can non-lumberjacks jack wood (with stones or weapons)

Lowering it to 1 would be too harsh on the lumberjacks maybe? Maybe not?
Maybe Steward can adjust it anyway

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->