Changes the Merchant's trait name (#425)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Changes the "Giza Blooded" trait's name to "Skilled Appraiser" and
"Giza-in-training" to "Merchant-in-training" (If anyone has better
recommendations for the name do tell me).
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Made sense when it was mainly a Human role, but since now any race can
be merchant it makes less sense for them to be "Giza Blooded", also
Sembia mentioned to change it.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->