MMB-on-turf now counts how many of each item there is (#412)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
The menu that comes up when middle-clicking a turf (i.e. a floor) now
gives you the numbers of each item present.
Additionally, attempting to take an item that gets deleted after you
middle clicked will no longer cause a runtime.

More technically, it gives a list of item names with a count of each
item that shares the same name.
For example, iron "sword"s and steel "sword"s, which share the same
name, will be grouped together.
The first corresponding item for each name is marked when you
middle-click, and is what you will take upon selection.
If an iron sword comes up first in the alt-RMB list for a pile with
steel swords mixed in, that's the one you will take.

![image](https://github.com/user-attachments/assets/f46dd577-6851-455b-82f7-2ce2b7fbaf75)

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
The turf middle-click function actually has a purpose now, the same
purpose it used to have. Player-to-player bulk commerce is now more
feasible, item stacks in chests, handcarts and sacks are more manageable
(farming!), and you don't have to scroll through an alt-RMB list of
items to find just one that you're looking for.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

---------

Co-authored-by: zoktiik <zoktiik@users.noreply.github.com>