Vile prayers no longer set you on fire (#390)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
This was problematic for
1. People setting themselves on fire to run into vampires
2. People purposefully committing suicide while being captured with this
3. People having legitimate prayers to inhumen gods with the vile words
included being punished

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->