Pixel Shifting - The HRP Aid. (#380)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

This PR adds pixel shifting, hotkeyed by default to Ctrl+Shift+movement
keys, and can be remapped (Note, old players will need to reset your
keybinds for the default hotkeys to be selected). The pixel shift is
limited to 16 pixels and resets to normal after moving. Pixel shifting
while wallpressed is disabled, as is pixel-shifting and then
wall-leaning. In order to share a turf with someone, have the person in
on the desired turf lie down and stand ontop of them, then you can both
pixel shifted into the desired positions.

## Why It's Good For The Game

Pixel shifting has been a wanted feature from the start, and this adds
it. Pixel shifting helps with HRP and especially while HRPing the
holding of hands.