Hoods can now be worn in the cloak slot, also red hood added to mage hood list (#337)
basically the title

it would be nice to have a covering that can go on the neck that doesn't
cover your tail - this does that! i also like the way the hood neck
thing looks on the neck but want a helmet with it 'cuz... why couldn't
you? was going to put it in the neck slot or mask slot like how you can
the kiffeyeh but felt it more fit in cloak slot given it can hide
identity