Restores the default "storkrieg combat music" (#404)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Quick PR at the request of Azarak and others to replace the older combat
music with the newer, better track.
This music plays when nothing else, such as your job or villain role,
overrides the default combat music.
The original commit was 7ffc2e3d07206e46724f836c06815bbfd4f6b827.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
This is a better track. People know it, people like it.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

Co-authored-by: zoktiik <zoktiik@users.noreply.github.com>