Butcher's loadout changes (#443)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Makes the Butcher spawn with the key to their shop, and makes the
loadout code a bit more compact.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
All roles spawn with keys to their place, so it makes sense to me for
them to start with a key too, also following what was done with the
Merchant's outfit I though it would be nice to unify the Butcher's
outfit for the thing both genders get.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->