Remove Maniac's Gender Lock. (#445)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Remove the gender lock for Maniac.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
uh, Even for the lore Implication I don't really get it and honestly
talking,
Troy problably is fine with playing a female character, I suppose. 
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

---------

Co-authored-by: Azarak <azarak10@gmail.com>