Changes Modern hairstyle, SLIGHTLY changes pighelm (#468)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
I am learning how to use dreammaker and make PRs for the first time and
thought two minor sprite edits that were consistently and constantly
bothering me would be a good way to learn. I tweaked a large number of
things on the "modern" hairstyle to make it work with the game's helmets
and to reduce the galaxybrain it had going on, and I very slightly
tweaked the pig's bascinet so the raised visor isn't inexplicably
shrinking between forward and backward sprites. This does not add it as
a new hairstyle purely because I believe the new version is too similar
to justify it.


![dreamseeker_BGxqaRlWr0](https://github.com/user-attachments/assets/28c1ac24-de92-4952-8f6c-696cfcf7f863)

I have been told it is customary tradition to write "Please speed-merge
my PR".
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Having hairstyles that look like Megamind in certain directions is BAD,
having hairstyles that fit into helmets is GOOD. Having visors that look
inexplicably smaller facing forwards than backwards is BAD, having
visors that actually look like you're looking at the front and back of
the visor is GOOD.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->