Reverts modern hair (#476)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
I realised what exactly
https://github.com/Rotwood-Vale/Ratwood-Keep/pull/468 was doing, and
this reverts the modern hair change

The proper solution to the problem that PR is tackling is a "dynamic
extension" framework for hair that is covered by helmets (One such that
exists on TG for example). But it would require more effort to bring it
here

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->