Don't Touch ME! A Anti ERP Option
Adds a toggle to the options menu that stops people from being able to use the ERP panel on you.
This will stop ERP half way thru if turned on part way thru.
Gives a notification in chat if someone has it on and where the option is (So they can find it even if they are a new player)
It is set to not allow it by default. You will have to ALLOW people to touch you.