Adds "Modern 2" (#487)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

In #468, alongside a slight tweak to the pig's helm I heavily changed
the Modern hairstyle once I saw that it's inexplicably bigger than other
hairstyles and disproportionately bobbleheaded. This was reverted in
shaved a hairstyle down to fit helmets but an actual resprite and
touch-up, I was told somebody actually complained about the change to
"their" hairstyle, so this keeps the version I fixed up because I think
it's a good sprite while keeping the old hydrocephalic version for that
one person's hydrocephalic character.

![dreamseeker_7dQHK4ZfwL](https://github.com/user-attachments/assets/23d3e9c8-07d4-4e13-9acf-87e01b9f37f9)
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

Having hairstyles that look like Megamind in certain directions is BAD,
but apparently someone likes them. Learning how to add whole new
hairstyles is GOOD. Keeps my improved version.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

---------

Co-authored-by: Azarak <azarak10@gmail.com>